% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augknt.R
\name{augknt}
\alias{augknt}
\title{augknt}
\usage{
augknt(boundary, nBspline, order)
}
\arguments{
\item{boundary}{boundary of knots}

\item{nBspline}{number of b splines}

\item{order}{order of b spline function, indicates the degree n-1 of piecewise polynomial function}
}
\description{
This function is used to find knots
}
\examples{
boundary = c(0.1, 0.5)
nBspline = 10
order = 3
print(augknt(boundary, nBspline, order))

}
