% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegressor.R
\name{getRegressor}
\alias{getRegressor}
\title{get Regressor}
\usage{
getRegressor(x, d, spconfig)
}
\arguments{
\item{x}{input data, vector x}

\item{d}{the index of which knots to look at}

\item{spconfig}{splines configuration}
}
\value{
a vector of regressor that will be passed to EM algorithm
}
\description{
This function preprocess the original data to get splines configuration and input data for b splines.
}
\examples{
x =  matrix(rnorm(2*(1000+8)),1000+8, 2)
d = rep(1,length(x[,1]))
knotBox = apply(x, 2, quantile, c(0.02,0.98), na.rm = TRUE)
nBspline = 10
order = 3
knots = apply(knotBox, 2, augknt, nBspline, order)
spconfig = list(order = 3, nBspline = 10, knots = knots,knotBox = knotBox)
regressor <- getRegressor(x[,1],d,spconfig)
}
