% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoeff.R
\name{plotcoeff}
\alias{plotcoeff}
\title{plotcoeff}
\usage{
plotcoeff(beta, knots, nBspline)
}
\arguments{
\item{beta}{optimal coefficients}

\item{knots}{knots that returned from getPreprocess function}

\item{nBspline}{the numbers of B splines that we need to fit model, also returned from getPreprocess function}
}
\description{
This function is to plot the coefficient
}
\examples{
positive.polynomial(0.2,x = 0,k = 3,end_knot = 2,d=0)
}
