% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_polynomial.R
\name{positive.polynomial}
\alias{positive.polynomial}
\title{positive.polynomial}
\usage{
positive.polynomial(t, x, k, end_knot, d = 0)
}
\arguments{
\item{t}{the input value}

\item{x}{knots value}

\item{k}{order}

\item{end_knot}{the largest value for knot}

\item{d}{the dth derivative if wanted}
}
\value{
a number of positive polynomial result
}
\description{
This function works as truncated polynomial function
}
\examples{
positive.polynomial(0.2,x = 0,k = 3,end_knot = 2,d=0)
}
