% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spval.R
\name{spval}
\alias{spval}
\title{divided difference}
\usage{
spval(x, knots)
}
\arguments{
\item{x}{input x value}

\item{knots}{knots}
}
\value{
regressor value
}
\description{
This function is to calculate single value of regressor
}
\examples{
knots = c(0,0,0,0.1)
input = 0.02
result = spval(input,knots)

}
