% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_EM.R
\name{glasso_EM}
\alias{glasso_EM}
\title{glasso_EM}
\usage{
glasso_EM(beta_init, A, B, ngroup, lambda, alpha2, ...)
}
\arguments{
\item{beta_init}{initial value}

\item{A}{sufficient statistics, matrix}

\item{B}{sufficient statistics}

\item{ngroup}{number of groups}

\item{lambda}{the LASSO penalty}

\item{alpha2}{the EM decomposition parameter, refer to the paper}

\item{...}{default parameters}
}
\value{
a list of whether success and the updated coefficients
}
\description{
This function is to implement the em algorithm for group lasso
}
\examples{
beta = matrix(c(1,2,3))
B = matrix(c(1,2,3))
A = matrix(rep(c(1,2,3),3),nrow = 3,byrow = TRUE)
ngroup = 3
lambda = 0.05
alpha2 = 0.05
out = glasso_EM(beta,A,B,ngroup,lambda,alpha2)

}
