% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineModel.R
\name{onlineModel}
\alias{onlineModel}
\title{onlineModel}
\usage{
onlineModel(y, x, D, L, lambda, gamma_init, alpha2_init, spconfig,
  spaTol_gamma, shrinkStepSize, moveSize, ...)
}
\arguments{
\item{y}{vector}

\item{x}{the reuslt from getPreprocess \code{x}}

\item{D}{columns of original X}

\item{L}{maximize lag number}

\item{lambda}{the LASSO penalty}

\item{gamma_init}{the initial value for gamma}

\item{alpha2_init}{the initial value of EM decomposition parameter, refer to the paper}

\item{spconfig}{spline configuration}

\item{spaTol_gamma}{favor smaller gamma within spaTol_c tolerance}

\item{shrinkStepSize}{adjust gamma before they get too crazy}

\item{moveSize}{multipler to move among channels}

\item{...}{default parameters,including
 \itemize{
  \item testSize: the number of obs to average prediction error, default is 50
  \item tol_idle_alpha2:increase alpha2 if it idles for a while
  \item safeShrink_gamma : adjust gamma before they get too crazy
  \item tol_all0times: if coefficents are all 0 for too long time, shrink gamma
  \item shrinkAlpha2: how to shrink alpha2
}}
}
\value{
a list including
 \itemize{
  \item preErr: historical predicted error, firt column indicates the first channel, etc.
  \item predict1 : historical predicted values, firt column indicates the first channel, etc.
  \item gamma_history : historical gamma channels, first channel is the smallest one and the third channel is the largest one.
  \item gamma_opt: optimized gamma, the second channel in gamma_history
  \item alpha_opt: optimized alpha
  \item beta_opt: best coefficient, the last row and the second channel related in beta_history
  \item beta_history: all stored coefficient.
}
}
\description{
This model is used to fit sequential non linear time series data. Detailed information could be find in the paper \url{http://jding.org/jie-uploads/2018/11/slant.pdf}
}
\details{
This function is for online streaming data
}
\seealso{
\code{\link{glasso_EM}} for implementation of EM algorithm; \code{\link{getPreprocess}};\code{\link{getRegressor}}
}
\keyword{linear}
\keyword{model}
\keyword{non}
\keyword{sequential}
\keyword{series}
\keyword{time}
