% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoeff.R
\name{plotcoeff}
\alias{plotcoeff}
\title{plotcoeff}
\usage{
plotcoeff(beta, knots, nBspline, order)
}
\arguments{
\item{beta}{optimal coefficients}

\item{knots}{first column of knots that returned from getPreprocess function, generated from original scale}

\item{nBspline}{the numbers of B splines that we need to fit model, also returned from getPreprocess function}

\item{order}{the order of b splines}
}
\description{
This function is to plot the coefficient
}
\examples{
beta = c(rep(1,10),rep(0,140),rep(-1,10))
knots = c(0.1,0.1,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.8,0.8,0.8)
nBspline = 10
order = 3
plotcoeff(beta,knots,nBspline,order)
}
