% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spval.R
\name{spval}
\alias{spval}
\title{divided difference}
\usage{
spval(t, x)
}
\arguments{
\item{t}{input x value}

\item{x}{knots}
}
\value{
regressor value
}
\description{
This function is to calculate single value of regressor
}
